// src/types/CSL.ts
export enum CameraVerticalAngle {
  WormsEye = "wormsEye",
  Low = "low",
  Eye = "eye",
  High = "high",
  Overhead = "overhead",
  BirdsEye = "birdsEye",
  TopDown = "topDown",
}


export enum ShotSize {
  ExtremeCloseUp = "extremeCloseUp",
  CloseUp = "closeUp",
  MediumCloseUp = "mediumCloseUp",
  MediumShot = "mediumShot",
  MediumLongShot = "mediumLongShot",
  FullShot = "fullShot",
  LongShot = "longShot",
  VeryLongShot = "veryLongShot",
  ExtremeLongShot = "extremeLongShot",
}


export enum SubjectView {
  Front = "front",
  Back = "back",
  Left = "left",
  Right = "right",
  ThreeQuarterFrontLeft = "threeQuarterFrontLeft",
  ThreeQuarterFrontRight = "threeQuarterFrontRight",
  ThreeQuarterBackLeft = "threeQuarterBackLeft",
  ThreeQuarterBackRight = "threeQuarterBackRight",
}
// ...
export interface Vector3 {
  x: number;
  y: number;
  z: number;
}

export interface Quaternion {
  x: number;
  y: number;
  z: number;
  w: number;
}

export interface EulerAngles {
  pitch: number;
  yaw: number;
  roll: number;
}

export interface CameraPose {
  position: Vector3;
  rotation: Quaternion | EulerAngles;
}

export interface CameraExtrinsics {
  pose: CameraPose;
  transformMatrix?: [
    [number, number, number, number],
    [number, number, number, number],
    [number, number, number, number],
    [number, number, number, number]
  ];
}

export interface CameraIntrinsics {
  focalLength?: number; // in mm
  fov?: number;
  aspectRatio?: number;
  sensorSize?: { width: number; height: number }; // in mm
}


export interface Target {
  id: string;
  description: string;
}

export interface SubjectFraming {
  position?: SubjectInFramePosition;
  dutchAngleScale?: Scale;
}

export interface SubjectAwareCameraConfig {
  type: "subjectAware";
  cameraAngle?: CameraVerticalAngle;
  shotSize?: ShotSize;
  subjectView?: SubjectView;
  subjectFraming?: SubjectFraming;
}

export interface NonSubjectAwareCameraConfig {
  type: "nonSubjectAware";
  extrinsics: CameraExtrinsics;
  intrinsics?: CameraIntrinsics;
  lookAt?: Vector3 | Target[];
}


export type CameraConfig = SubjectAwareCameraConfig | NonSubjectAwareCameraConfig;
// ...
export interface SinglePointConstraint {
  type: "singlePoint";
  time: number; // in seconds
  config: CameraConfig;
  weight?: number; // Importance in optimization (0.0 to 1.0)
}

export interface IntervalConstraint {
  type: "interval";
  startTime: number;
  endTime: number;
  lossFunction: LossFunction;
  weight?: number;
}

export interface LossFunction {
  type: LossFunctionType;
  parameters: Record<string, unknown>;
}

export enum LossFunctionType {
  PedestalMovement = "pedestalMovement",
  DollyMovement = "dollyMovement",
  TruckMovement = "truckMovement",
  PanMovement = "panMovement",
  TiltMovement = "tiltMovement",
  ArcMovement = "arcMovement",
  Static = "Static",
  FollowMovement = "followMovement",

  FramingPosition = "framingPosition",
  ShotSize = "shotSize",
  SubjectView = "subjectView",

  // General losses
  Collision = "collision",
  Smoothness = "smoothness",
  MinPath = "minPath",
}

export interface SectionSolverOutput {
  initKeyframes: SinglePointConstraint[];
  constraints: (SinglePointConstraint | IntervalConstraint)[];
}

export interface TimelineSolverOutput {
  sections: SectionSolverOutput[];
}

export type TimelineSegment =
  | {
    kind: "interval";
    startTime: number;
    endTime: number;
    lossFunctions: LossFunction[];
  }
  | {
    kind: "point";
    time: number;
    lossFunctions: LossFunction[];
  };

export type NormalizedTimeline = TimelineSegment[];